+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 [Unreleased]
 
-- No unreleased changes
+- New card "Cliffcrush Giant" (thanks @Zayberex)
+- New card "Cantankerous Giant" (thanks @Zayberex)
+- New card "Factory Shell Q" (thanks @Zayberex)
+- New card "Rain of Arrows" (thanks @jyotiskaghosh)
+- New card "Comet Missile" (thanks @jyotiskaghosh)
+- New card "Crisis Boulder" (thanks @jyotiskaghosh)
+- New card "Grave Worm Q" (thanks @Zayberex)
+- New card "Tentacle Worm" (thanks @Zayberex)
+- New card "Aeropica" (thanks @Zayberex)
+- New card "Zepimeteus" (thanks @Zayberex)
+- New card "Promephius Q" (thanks @Zayberex and @xrenari)
 
 ## 23/12/2023
 
@@ -11,6 +21,8 @@
 - Fixed typo in chat message for "Onslaughter Triceps" (thanks @Zayberex)
 - Fixed a formatting issue wiht the chat message sent by "Soul Gulp" (thanks @Zayberex)
 - Updated "Daidalos, General Of Fury"'s effect to not break shields if the creature chosen to be destroyed is itself (thanks @Zayberex)
+- New card "Lightning Grass" (thanks @Zayberex)
+- New card "Razorpine Tree" (thanks @Zayberex)
 - New card "Q-tronic Gargantua" (thanks @Zayberex)
 - New card "Steam Star" (thanks @Zayberex)
 - New card "Ripple Lotus Q" (thanks @Zayberex)